phosphor-ipmi-flash: Fix all C++ warnings

+++ bmc/firmware-handler/test/firmware_state_notyetstarted_tarball_unittest.cpp
@@ -84,8 +84,8 @@ class FirmwareHandlerNotYetStartedUbitTest : public ::testing::Test
     TriggerMock* updateMockPtr;
 
     std::uint16_t session = 1;
-    std::uint16_t flags =
-        blobs::OpenFlags::write | FirmwareFlags::UpdateFlags::ipmi;
+    std::uint16_t flags = static_cast<std::uint16_t>(blobs::OpenFlags::write) |
+                          FirmwareFlags::UpdateFlags::ipmi;
 };
 
 TEST_F(FirmwareHandlerNotYetStartedUbitTest,
